+++ app/src/main/java/com/techcourse/controller/RegisterViewController.java
@@ -2,7 +2,7 @@ package com.techcourse.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
+import com.interface21.webmvc.servlet.mvc.controller.Controller;
 
 public class RegisterViewController implements Controller {
