+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/controller/ForwardController.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.asis;
+package com.interface21.webmvc.servlet.mvc.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
